pyproject.toml
[tool.poetry]
name = "drf-boilerplate"
version = "0.1.0"
description = "A boilerplate for DRF applications"
authors = ["carvalhocaio <caiocarvalho.py@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.3"
django = "4.2"
python-decouple = "^3.8"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.5"
pre-commit = "^3.6.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.7"
black = "^23.11.0"
ipdb = "^0.13.13"
pytest-django = "^4.7.0"
pytest-cov = "^4.1.0"
pytest-factoryboy = "^2.6.0"
faker = "^20.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
